chore(accordion): 🏷️  update type declaration for prop generation
