+++ src/accordion/AccordionState.ts
@@ -1,32 +1,18 @@
 import * as React from "react";
 import { useControllableState } from "@chakra-ui/hooks";
 import { CompositeInitialState, useCompositeState } from "reakit";
+
 import {
-  AccordionState,
   SelectedIdPair,
+  AccordionState,
   AccordionActions,
   AccordionReturns,
-  MultiOverloadReturn,
-  SingleOverloadReturn,
   AccordionPropsUnion,
+  MultiOverloadReturn,
   MultiAccordionProps,
+  SingleOverloadReturn,
   SingleAccordionProps,
 } from "./types";
-
-export type AccordionInitialState = Partial<
-  AccordionPropsUnion & CompositeInitialState
->;
-export type AccordionInitialStateSingle = Partial<
-  SingleAccordionProps & CompositeInitialState
->;
-export type AccordionInitialStateMulti = Partial<
-  MultiAccordionProps & CompositeInitialState
->;
-
-export type AccordionStateReturn = AccordionActions &
-  AccordionState &
-  AccordionReturns &
-  SelectedIdPair;
 
 export function useAccordionState(
   props: AccordionInitialStateSingle,
@@ -156,3 +142,18 @@ export function useAccordionState(
         ...common,
       };
 }
+
+export type AccordionInitialState = Partial<
+  AccordionPropsUnion & CompositeInitialState
+>;
+export type AccordionInitialStateSingle = Partial<
+  SingleAccordionProps & CompositeInitialState
+>;
+export type AccordionInitialStateMulti = Partial<
+  MultiAccordionProps & CompositeInitialState
+>;
+
+export type AccordionStateReturn = AccordionState &
+  AccordionActions &
+  AccordionReturns &
+  SelectedIdPair;
